requirements.txt
flask==3.1.1
flask-restx==1.3.0
requests==2.32.3
python-dotenv==1.1.0
gunicorn==23.0.0
flask-cors==5.0.1
